Create LeetCode_Problem_No_27_Remove_Element.java
Java Solution of Leet Code Problem Number 27 - Remove Element

Problem Description - Given an integer array nums and an integer val, remove all occurrences of val in nums in-place. The order of the elements may be changed. Then return the number of elements in nums which are not equal to val.